pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "toroidal-sneks-submission"
version = "1.0.0"
description = "Template repository for submissions to sneks.dev"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache Software License"}
authors = [
    {name = "Nathaniel Schaaf", email = "nathaniel.schaaf@gmail.com"}
]
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "toroidal-sneks>=1.0.3"
]

[project.scripts]
validate = "scripts:validate"
run = "scripts:run"

[tool.isort]
profile = "black"
